# docs/site/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Memobase API",
		"summary": "APIs for Memobase, a user memory system for LLM Apps",
		"version": "0.0.29"
	},
	"servers": [
		{
			"url": "https://api.memobase.dev"
		},
		{
			"url": "https://api.memboase.cn"
		}
	],
	"paths": {
...
		"/api/v1/project/profile_config": {
			"get": {
				"tags": [
					"project"
				],
				"summary": "Get Project Profile Config String",
				"operationId": "get_project_profile_config_string_api_v1_project_profile_config_get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileConfigDataResponse"
								}
							}
						}
					}
				},
				"x-code-samples": [
					{
						"lang": "Python",
						"source": "# To use the Python SDK, install the package:\n# pip install memobase\n\nfrom memobase import Memobase\n\nmemobase = Memobase(project_url='PROJECT_URL', api_key='PROJECT_TOKEN')\n\nconfig = memobase.get_config()\n",
						"label": "Python"
					},
					{
						"lang": "JavaScript",
						"source": "// To use the JavaScript SDK, install the package:\n// npm install @memobase/memobase\n\nimport { MemoBaseClient } from '@memobase/memobase';\n\nconst client = new MemoBaseClient(process.env.MEMOBASE_PROJECT_URL, process.env.MEMOBASE_API_KEY);\n\nconst config = await client.getConfig();\n",
						"label": "JavaScript"
					}
				]
			},
			"post": {
				"tags": [
					"project"
				],
				"summary": "Update Project Profile Config",
				"operationId": "update_project_profile_config_api_v1_project_profile_config_post",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProfileConfigData",
								"description": "The profile config to update"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				},
				"x-code-samples": [
					{
						"lang": "Python",
						"source": "# To use the Python SDK, install the package:\n# pip install memobase\n\nfrom memobase import Memobase\n\nmemobase = Memobase(project_url='PROJECT_URL', api_key='PROJECT_TOKEN')\n\nmemobase.update_config('your_profile_config')\n",
						"label": "Python"
					},
					{
						"lang": "JavaScript",
						"source": "// To use the JavaScript SDK, install the package:\n// npm install @memobase/memobase\n\nimport { MemoBaseClient } from '@memobase/memobase';\n\nconst client = new MemoBaseClient(process.env.MEMOBASE_PROJECT_URL, process.env.MEMOBASE_API_KEY);\n\nawait client.updateConfig('your_profile_config');\n",
						"label": "JavaScript"
					}
				]
			}
		},
...
		"/api/v1/users": {
			"post": {
				"tags": [
					"user"
				],
				"summary": "Create User",
				"description": "Create a new user with additional data",
				"operationId": "create_user_api_v1_users_post",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserData",
								"description": "User data for creating a new user"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IdResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				},
				"x-code-samples": [
					{
						"lang": "Python",
						"source": "# To use the Python SDK, install the package:\n# pip install memobase\n\nfrom memobase import Memobase\n\nclient = Memobase(project_url='PROJECT_URL', api_key='PROJECT_TOKEN')\n\nuid = client.add_user({\"ANY\": \"DATA\"})\n",
						"label": "Python"
					},
					{
						"lang": "JavaScript",
						"source": "// To use the JavaScript SDK, install the package:\n// npm install @memobase/memobase\n\nimport { MemoBaseClient } from '@memobase/memobase';\n\nconst client = new MemoBaseClient(process.env.MEMOBASE_PROJECT_URL, process.env.MEMOBASE_API_KEY);\n\nconst userId = await client.addUser({ANY: \"DATA\"});\n",
						"label": "JavaScript"
					},
					{
						"lang": "Go",
						"source": "// To use the Go SDK, install the package:\n// go get github.com/memodb-io/memobase/src/client/memobase-go@latest\n\nimport (\n    \"github.com/memodb-io/memobase/src/client/memobase-go/core\"\n    \"github.com/google/uuid\"\n)\n\nprojectURL := \"YOUR_PROJECT_URL\"\napiKey := \"YOUR_API_KEY\"\nclient, err := core.NewMemoBaseClient(projectURL, apiKey)\nif err != nil {\n    panic(err)\n}\n\n// Generate a UUID for the user\nuserID := uuid.New().String()\n\n// Create user with some data\ndata := map[string]interface{}{\"ANY\": \"DATA\"}\nresultID, err := client.AddUser(data, userID)\nif err != nil {\n    panic(err)\n}\n",
						"label": "Go"
					}
				]
			}
		},
...
		"/api/v1/users/blobs/{user_id}/{blob_type}": {
			"get": {
				"tags": [
					"user"
				],
				"summary": "Get User All Blobs",
				"operationId": "get_user_all_blobs_api_v1_users_blobs__user_id___blob_type__get",
				"parameters": [
					{
						"name": "user_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"description": "The ID of the user to fetch blobs for",
							"title": "User Id"
						},
						"description": "The ID of the user to fetch blobs for"
					},
					{
						"name": "blob_type",
						"in": "path",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/BlobType",
							"description": "The type of blobs to retrieve"
						},
						"description": "The type of blobs to retrieve"
					},
					{
						"name": "page",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"description": "Page number for pagination, starting from 0",
							"default": 0,
							"title": "Page"
						},
						"description": "Page number for pagination, starting from 0"
					},
					{
						"name": "page_size",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"description": "Number of items per page, default is 10",
							"default": 10,
							"title": "Page Size"
						},
						"description": "Number of items per page, default is 10"
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IdsResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				},
				"x-code-samples": [
					{
						"lang": "Python",
						"source": "# To use the Python SDK, install the package:\n# pip install memobase\n\nfrom memobase import Memobase\nfrom memobase.core.types import BlobType\n\nmemobase = Memobase(project_url='PROJECT_URL', api_key='PROJECT_TOKEN')\n\nuser = memobase.get_user('user_id')\nblobs = user.get_all(BlobType.CHAT)\n",
						"label": "Python"
					},
					{
						"lang": "JavaScript",
						"source": "// To use the JavaScript SDK, install the package:\n// npm install @memobase/memobase\n\nimport { MemoBaseClient, BlobType } from '@memobase/memobase';\n\nconst client = new MemoBaseClient(process.env.MEMOBASE_PROJECT_URL, process.env.MEMOBASE_API_KEY);\n\nconst user = client.getUser('user_id');\nconst blobs = await user.getAll(BlobType.Enum.chat);\n",
						"label": "JavaScript"
					},
					{
						"lang": "Go",
						"source": "// To use the Go SDK, install the package:\n// go get github.com/memodb-io/memobase/src/client/memobase-go@latest\n\nimport (\n    \"github.com/memodb-io/memobase/src/client/memobase-go/core\"\n    \"github.com/memodb-io/memobase/src/client/memobase-go/blob\"\n)\n\nprojectURL := \"YOUR_PROJECT_URL\"\napiKey := \"YOUR_API_KEY\"\nclient, err := core.NewMemoBaseClient(projectURL, apiKey)\nif err != nil {\n    panic(err)\n}\n\n// Get user\nuser, err := client.GetUser(userID)\nif err != nil {\n    panic(err)\n}\n\n// Get all blobs\nblobs, err := user.GetAll(blob.ChatType)\nif err != nil {\n    panic(err)\n}\n",
						"label": "Go"
					}
				]
			}
		},
		"/api/v1/blobs/insert/{user_id}": {
			"post": {
				"tags": [
					"blob"
				],
				"summary": "Insert Blob",
				"operationId": "insert_blob_api_v1_blobs_insert__user_id__post",
				"parameters": [
					{
						"name": "user_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"description": "The ID of the user to insert the blob for",
							"title": "User Id"
						},
						"description": "The ID of the user to insert the blob for"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BlobData",
								"description": "The blob data to insert"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IdResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				},
				"x-code-samples": [
					{
						"lang": "Python",
						"source": "# To use the Python SDK, install the package:\n# pip install memobase\n\nfrom memobase import Memobase\nfrom memobase import ChatBlob\n\nclient = Memobase(project_url='PROJECT_URL', api_key='PROJECT_TOKEN')\n\nb = ChatBlob(messages=[\n    {\n        \"role\": \"user\",\n        \"content\": \"Hi, I'm here again\"\n    },\n    {\n        \"role\": \"assistant\",\n        \"content\": \"Hi, Gus! How can I help you?\"\n    }\n])\nu = client.get_user(uid)\nbid = u.insert(b)\n",
						"label": "Python"
					},
					{
						"lang": "JavaScript",
						"source": "// To use the JavaScript SDK, install the package:\n// npm install @memobase/memobase\n\nimport { MemoBaseClient, Blob, BlobType } from '@memobase/memobase';\n\nconst client = new MemoBaseClient(process.env.MEMOBASE_PROJECT_URL, process.env.MEMOBASE_API_KEY);\nconst user = await client.getUser(userId);\n\nconst blobId = await user.insert(Blob.parse({\n  type: BlobType.Enum.chat,\n  messages: [\n    {\n      role: 'user',\n      content: 'Hi, I'm here again'\n    },\n    {\n      role: 'assistant',\n      content: 'Hi, Gus! How can I help you?'\n    }\n  ]\n}));\n",
						"label": "JavaScript"
					},
					{
						"lang": "Go",
						"source": "// To use the Go SDK, install the package:\n// go get github.com/memodb-io/memobase/src/client/memobase-go@latest\n\nimport (\n    \"github.com/memodb-io/memobase/src/client/memobase-go/core\"\n    \"github.com/memodb-io/memobase/src/client/memobase-go/blob\"\n)\n\nprojectURL := \"YOUR_PROJECT_URL\"\napiKey := \"YOUR_API_KEY\"\nclient, err := core.NewMemoBaseClient(projectURL, apiKey)\nif err != nil {\n    panic(err)\n}\n\n// Get user\nuser, err := client.GetUser(userID)\nif err != nil {\n    panic(err)\n}\n\n// Create chat blob\nchatBlob := &blob.ChatBlob{\n    BaseBlob: blob.BaseBlob{\n        Type: blob.ChatType,\n    },\n    Messages: []blob.OpenAICompatibleMessage{\n        {\n            Role:    \"user\",\n            Content: \"Hi, I'm here again\",\n        },\n        {\n            Role:    \"assistant\",\n            Content: \"Hi, Gus! How can I help you?\",\n        },\n    },\n}\n\n// Insert blob\nblobID, err := user.Insert(chatBlob)\nif err != nil {\n    panic(err)\n}\n",
						"label": "Go"
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
			"BaseResponse": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"type": "object"
							},
							{
								"type": "null"
							}
						],
						"title": "Data",
						"description": "Response data payload"
					},
					"errno": {
						"$ref": "#/components/schemas/CODE",
						"description": "Error code, 0 means success",
						"default": 0
					},
					"errmsg": {
						"type": "string",
						"title": "Errmsg",
						"description": "Error message, empty when success",
						"default": ""
					}
				},
				"type": "object",
				"title": "BaseResponse"
			},
...
			"BlobType": {
				"type": "string",
				"enum": [
					"chat",
					"doc",
					"image",
					"code",
					"transcript"
				],
				"title": "BlobType"
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"IdResponse": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/IdData"
							},
							{
								"type": "null"
							}
						],
						"description": "Response containing a single ID"
					},
					"errno": {
						"$ref": "#/components/schemas/CODE",
						"description": "Error code, 0 means success",
						"default": 0
					},
					"errmsg": {
						"type": "string",
						"title": "Errmsg",
						"description": "Error message, empty when success",
						"default": ""
					}
				},
				"type": "object",
				"title": "IdResponse"
			},
...
			"ProfileConfigData": {
				"properties": {
					"profile_config": {
						"type": "string",
						"title": "Profile Config",
						"description": "Profile config string"
					}
				},
				"type": "object",
				"required": [
					"profile_config"
				],
				"title": "ProfileConfigData"
			},
			"ProfileConfigDataResponse": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ProfileConfigData"
							},
							{
								"type": "null"
							}
						],
						"description": "Response containing profile config data"
					},
					"errno": {
						"$ref": "#/components/schemas/CODE",
						"description": "Error code, 0 means success",
						"default": 0
					},
					"errmsg": {
						"type": "string",
						"title": "Errmsg",
						"description": "Error message, empty when success",
						"default": ""
					}
				},
				"type": "object",
				"title": "ProfileConfigDataResponse"
			},
...
		},
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"security": [
		{
			"BearerAuth": []
		}
	]
}